requirements.txt
biopython
scipy
numpy
tqdm